requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
gino==1.0.1
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
SQLAlchemy==1.3.24
urllib3==1.26.9
yarl==1.7.2
